+++ Application/Scene/Sweep.h
@@ -25,7 +25,7 @@ class CSweep : public CMesh
     DEFINE_INPUT(CSweepPathInfo *, CrossSection) { MarkDirty(); }
     DEFINE_INPUT(float, Azimuth) { MarkDirty(); }
     DEFINE_INPUT(float, Twist) { MarkDirty(); }
-    DEFINE_INPUT(int, Join) { MarkDirty(); }
+    DEFINE_INPUT(float, Join) { MarkDirty(); }
 
 public:
     DECLARE_META_CLASS(CSweep, CMesh);
